Add new version of gate functions not to specify bit masks explicitly
The original implementation of gate functions requres qubit_masks and
index_masks for their arguments. However, this will not be accepted if
we implement qudit (NOT qubit) gate operations in the future beacuse
some (bit) operations cannot be applied in non-binary representation of
numeral systems.

In this implementation, the macro KET_USE_BIT_MASKS_EXPLICITLY
determines if bit masks are used explicitly or not. The reason why the
original implementation is still placed is performance.